Refactoring: move container to testhelpers package
