+++ Makefile
@@ -1,5 +1,5 @@
 lint:
-	golangci-lint run ./config/... ./database/... ./hasura/... ./node/... ./prometheus/... ./tools/... ./tzkt/...
+	golangci-lint run ./config/... ./database/... ./hasura/... ./node/... ./prometheus/... ./testhelpers/... ./tools/... ./tzkt/...
 
 test:
-	go test ./config/... ./database/... ./hasura/... ./node/... ./prometheus/... ./tools/... ./tzkt/...
+	go test ./config/... ./database/... ./hasura/... ./node/... ./prometheus/... ./testhelpers/... ./tools/... ./tzkt/...
