+++ config/go.mod
@@ -1,11 +1,11 @@
 module github.com/dipdup-io/go-lib/config
 
-go 1.25.0
+go 1.26.4
 
 require (
 	github.com/go-playground/validator/v10 v10.30.1
 	github.com/pkg/errors v0.9.1
-	golang.org/x/text v0.34.0
+	golang.org/x/text v0.37.0
 	gopkg.in/yaml.v3 v3.0.1
 )
 
@@ -14,6 +14,6 @@ require (
 	github.com/go-playground/locales v0.14.1 // indirect
 	github.com/go-playground/universal-translator v0.18.1 // indirect
 	github.com/leodido/go-urn v1.4.0 // indirect
-	golang.org/x/crypto v0.46.0 // indirect
-	golang.org/x/sys v0.39.0 // indirect
+	golang.org/x/crypto v0.52.0 // indirect
+	golang.org/x/sys v0.45.0 // indirect
 )
