+++ go.work
@@ -1,4 +1,4 @@
-go 1.25.7
+go 1.26.4
 
 use (
 	./config
@@ -6,6 +6,7 @@ use (
 	./hasura
 	./node
 	./prometheus
+	./testhelpers
 	./tools
 	./tzkt
 )
