+++ node/go.mod
@@ -1,6 +1,6 @@
 module github.com/dipdup-io/go-lib/node
 
-go 1.25.7
+go 1.26.4
 
 require (
 	github.com/json-iterator/go v1.1.12
@@ -13,5 +13,5 @@ require (
 	github.com/mattn/go-isatty v0.0.19 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
-	golang.org/x/sys v0.12.0 // indirect
+	golang.org/x/sys v0.44.0 // indirect
 )
