+++ prometheus/go.mod
@@ -1,9 +1,9 @@
 module github.com/dipdup-io/go-lib/prometheus
 
-go 1.25.7
+go 1.26.4
 
 require (
-	github.com/dipdup-io/go-lib/config v0.0.0-00010101000000-000000000000
+	github.com/dipdup-io/go-lib/config v1.0.0
 	github.com/prometheus/client_golang v1.23.2
 )
 
@@ -22,11 +22,9 @@ require (
 	github.com/prometheus/common v0.66.1 // indirect
 	github.com/prometheus/procfs v0.16.1 // indirect
 	go.yaml.in/yaml/v2 v2.4.2 // indirect
-	golang.org/x/crypto v0.46.0 // indirect
-	golang.org/x/sys v0.39.0 // indirect
-	golang.org/x/text v0.34.0 // indirect
+	golang.org/x/crypto v0.52.0 // indirect
+	golang.org/x/sys v0.45.0 // indirect
+	golang.org/x/text v0.37.0 // indirect
 	google.golang.org/protobuf v1.36.8 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
-
-replace github.com/dipdup-io/go-lib/config => ../config
