+++ tzkt/go.mod
@@ -1,9 +1,9 @@
 module github.com/dipdup-io/go-lib/tzkt
 
-go 1.25.7
+go 1.26.4
 
 require (
-	github.com/dipdup-io/go-lib/tools v0.0.0-00010101000000-000000000000
+	github.com/dipdup-io/go-lib/tools v1.0.0
 	github.com/gorilla/websocket v1.5.3
 	github.com/json-iterator/go v1.1.12
 	github.com/pkg/errors v0.9.1
@@ -24,5 +24,3 @@ require (
 	golang.org/x/sys v0.41.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
-
-replace github.com/dipdup-io/go-lib/tools => ../tools
